+++ src/conf/mod.rs
@@ -5,16 +5,6 @@ pub use parse::CmdOptConf;
 mod parse;
 
 impl CmdOptConf {
-    /*
-    pub fn base_dir(&self) -> String {
-        for o in self.opt_uc_x.iter() {
-            if let OptUcXParam::BaseDir(s) = o {
-                return s.clone();
-            }
-        }
-        String::new()
-    }
-    */
     pub fn is_opt_uc_x_help(&self) -> bool {
         for o in self.opt_uc_x.iter() {
             if let OptUcXParam::Help = o {
@@ -73,5 +63,28 @@ impl EnvConf {
 impl std::default::Default for EnvConf {
     fn default() -> EnvConf {
         EnvConf::new()
+    }
+}
+
+impl<IKV, K, V> From<IKV> for EnvConf
+where
+    IKV: IntoIterator<Item = (K, V)>,
+    K: AsRef<std::ffi::OsStr>,
+    V: AsRef<std::ffi::OsStr>,
+{
+    fn from(ary: IKV) -> Self {
+        let mut r = Self::new();
+        for a in ary {
+            match a.0.as_ref().to_string_lossy().to_string().as_str() {
+                "AKI_RESORT_COLOR_SEQ_ST" => {
+                    r.color_seq_start = a.1.as_ref().to_string_lossy().to_string();
+                }
+                "AKI_RESORT_COLOR_SEQ_ED" => {
+                    r.color_seq_end = a.1.as_ref().to_string_lossy().to_string();
+                }
+                _ => (),
+            }
+        }
+        r
     }
 }